# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Messenger API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  clerkId: string;
  username: string;
  displayName: string;
  avatarUrl?: string | null;
  bio?: string | null;
  isOnline: boolean;
  lastSeen?: string | null;
  createdAt: string;
}

export interface UpdateUserBody {
  username?: string;
  displayName?: string;
  bio?: string | null;
  avatarUrl?: string | null;
}

export type ChatType = (typeof ChatType)[keyof typeof ChatType];

export const ChatType = {
  direct: "direct",
  group: "group",
} as const;

export interface Chat {
  id: number;
  type: ChatType;
  name?: string | null;
  avatarUrl?: string | null;
  createdAt: string;
  createdBy: number;
}

export type ChatWithDetailsType =
  (typeof ChatWithDetailsType)[keyof typeof ChatWithDetailsType];

export const ChatWithDetailsType = {
  direct: "direct",
  group: "group",
} as const;

export type MessageType = (typeof MessageType)[keyof typeof MessageType];

export const MessageType = {
  text: "text",
  image: "image",
  file: "file",
  system: "system",
} as const;

export type MessageReactions = { [key: string]: number[] };

export interface Message {
  id: number;
  chatId: number;
  senderId: number;
  sender: User;
  content?: string | null;
  type: MessageType;
  replyToId?: number | null;
  replyTo?: Message | null;
  reactions: MessageReactions;
  isEdited: boolean;
  isDeleted: boolean;
  readBy: number[];
  createdAt: string;
  updatedAt: string;
}

export interface ChatWithDetails {
  id: number;
  type: ChatWithDetailsType;
  name?: string | null;
  avatarUrl?: string | null;
  createdAt: string;
  createdBy: number;
  lastMessage?: Message | null;
  members: User[];
  unreadCount: number;
}

export type CreateChatBodyType =
  (typeof CreateChatBodyType)[keyof typeof CreateChatBodyType];

export const CreateChatBodyType = {
  direct: "direct",
  group: "group",
} as const;

export interface CreateChatBody {
  type: CreateChatBodyType;
  name?: string | null;
  memberIds: number[];
}

export interface AddMemberBody {
  userId: number;
}

export type SendMessageBodyType =
  (typeof SendMessageBodyType)[keyof typeof SendMessageBodyType];

export const SendMessageBodyType = {
  text: "text",
  image: "image",
  file: "file",
} as const;

export interface SendMessageBody {
  content: string;
  type?: SendMessageBodyType;
  replyToId?: number | null;
}

export interface EditMessageBody {
  content: string;
}

export interface ReactBody {
  emoji: string;
}

export interface ChatStats {
  totalChats: number;
  totalMessages: number;
  totalUnread: number;
  activeChats: number;
}

export type SearchUsersParams = {
  q: string;
};

export type GetMessagesParams = {
  /**
   * Message ID cursor for pagination
   */
  before?: number;
  limit?: number;
};