# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Messenger API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User operations
  - name: chats
    description: Chat operations
  - name: messages
    description: Message operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username
      security:
        - clerkAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /users/online:
    get:
      operationId: getOnlineUsers
      tags: [users]
      summary: Get list of online users
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Online users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /chats:
    get:
      operationId: getChats
      tags: [chats]
      summary: Get all chats for current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatWithDetails"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create a new chat (direct or group)
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatBody"
      responses:
        "201":
          description: Created chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatWithDetails"

  /chats/{chatId}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get a single chat by ID
      security:
        - clerkAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatWithDetails"

  /chats/{chatId}/members:
    get:
      operationId: getChatMembers
      tags: [chats]
      summary: Get members of a chat
      security:
        - clerkAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of chat members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: addChatMember
      tags: [chats]
      summary: Add a member to a group chat
      security:
        - clerkAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberBody"
      responses:
        "200":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /chats/{chatId}/messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages in a chat
      security:
        - clerkAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: integer
          description: Message ID cursor for pagination
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a chat
      security:
        - clerkAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /chats/{chatId}/messages/{messageId}:
    put:
      operationId: editMessage
      tags: [messages]
      summary: Edit a message
      security:
        - clerkAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditMessageBody"
      responses:
        "200":
          description: Edited message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      security:
        - clerkAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /chats/{chatId}/messages/{messageId}/react:
    post:
      operationId: reactToMessage
      tags: [messages]
      summary: Add or toggle emoji reaction to a message
      security:
        - clerkAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactBody"
      responses:
        "200":
          description: Updated reactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /chats/{chatId}/messages/{messageId}/read:
    post:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark a message as read
      security:
        - clerkAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read

  /chats/stats:
    get:
      operationId: getChatStats
      tags: [chats]
      summary: Get chat stats for current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatStats"

components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        isOnline:
          type: boolean
        lastSeen:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - username
        - displayName
        - isOnline
        - createdAt

    UpdateUserBody:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true

    Chat:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [direct, group]
        name:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
      required:
        - id
        - type
        - createdAt
        - createdBy

    ChatWithDetails:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [direct, group]
        name:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
        lastMessage:
          $ref: "#/components/schemas/Message"
          nullable: true
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        unreadCount:
          type: integer
      required:
        - id
        - type
        - createdAt
        - createdBy
        - members
        - unreadCount

    CreateChatBody:
      type: object
      properties:
        type:
          type: string
          enum: [direct, group]
        name:
          type: string
          nullable: true
        memberIds:
          type: array
          items:
            type: integer
      required:
        - type
        - memberIds

    AddMemberBody:
      type: object
      properties:
        userId:
          type: integer
      required:
        - userId

    Message:
      type: object
      properties:
        id:
          type: integer
        chatId:
          type: integer
        senderId:
          type: integer
        sender:
          $ref: "#/components/schemas/User"
        content:
          type: string
          nullable: true
        type:
          type: string
          enum: [text, image, file, system]
        replyToId:
          type: integer
          nullable: true
        replyTo:
          $ref: "#/components/schemas/Message"
          nullable: true
        reactions:
          type: object
          additionalProperties:
            type: array
            items:
              type: integer
        isEdited:
          type: boolean
        isDeleted:
          type: boolean
        readBy:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - chatId
        - senderId
        - sender
        - type
        - reactions
        - isEdited
        - isDeleted
        - readBy
        - createdAt
        - updatedAt

    SendMessageBody:
      type: object
      properties:
        content:
          type: string
        type:
          type: string
          enum: [text, image, file]
          default: text
        replyToId:
          type: integer
          nullable: true
      required:
        - content

    EditMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    ReactBody:
      type: object
      properties:
        emoji:
          type: string
      required:
        - emoji

    ChatStats:
      type: object
      properties:
        totalChats:
          type: integer
        totalMessages:
          type: integer
        totalUnread:
          type: integer
        activeChats:
          type: integer
      required:
        - totalChats
        - totalMessages
        - totalUnread
        - activeChats